Alteration of Mongrel.run for Mongrel based routing
With the passing of the :map option Mongrel.run will handle the passing of a
Hash or URLMap rather than a standard rack app. The mapping provided by the
passed object will be used to register uris with the mongrel instance.

Hashes should only have absolute paths for keys and apps for values.

URLMaps will be filtered if the :Host options is specified, or the mapping's
host is set.

darcs-hash:20080311075137-f4dbf-59347707f098e5203c7eae6918050e31911606cb.gz