Fix that Request assumes form-data even when Content-Type says otherwise
Fixes cases where accessing Request#params causes the body to be read and
processed as form-data improperly. For example, consider the following
request:

    PUT /foo/bar?baz=bizzle HTTP/1.1
    Content-Type: text/plain

    This is not form-data.

When Rack::Request gets ahold of the corresponding environment, and the
application attempts to access the "baz" query string param, the body is read
and interpreted as form-data. If the body is an IOish object, this causes the
offset to be forwarded to the end of file.

The patch prevents the Request#POST method from going into the body unless the
media type is application/x-www-form-urlencoded, multipart/form-data, or not
specified.

While here, added a few unrelated helper methods to Request that I've found
particularly useful in Rack apps: #content_length, #head?, and #content_charset.

Tests and doc included for all changes.

darcs-hash:20080315135144-8bfab-f12391851f188ab1467b7e3d825f1e196058210d.gz