Rack::Etag should not set ETag if Last-Modified header is sent.
The HTTP specification says that if both Last-Modified and ETag are sent, both should be matched in other to check if a request is fresh or not. If Rack always set ETag, even if Last-Modified is sent, checking the freshness according to specification becomes non trivial, since I will always need the digested body response.

Signed-off-by: raggi <jftucker@gmail.com>