Byte-range support for File class.
Allows Rack to support byte-range requests (via
the HTTP 1.1 "Range:" header) for static files,
even when sendfile is not being used.

Conforms to RFC 2616 sec. 14.35 _except_ that
multiple byte-ranges are not supported yet.
(They're parsed correctly, but the response body
would need to be a MIME multipart.)

Tested in Ruby 1.8.7 on Mac OS X 10.6.4.