Allow the session to be skipped.
This will not send a cookie back nor change the session state.

The :defer option did not send the cookie back but did change the session
state in the backend.

This is useful for assets requests that still go through the rack stack
but do not want to cause any change in the session (for example accidentally expiring flash messages).