Set the gzip MTIME header to 0 when no Last-Modified header is available. #577
Often strong ETag validators are set on response headers. For instance with Rack::ETag, the ETag is a digest of the response body.
It is a waste of bandwidth to set the current time in the gzip header, because it causes all ETags to expire every second.
Also, most HTTP clients automatically inflate the response body. The MTIME of the gzip header is not used anyway.

The argument to generate the ETag based of the uncompressed body is invalid, unless weak ETag validators are being used.