Fix webrick handling of Transfer-encoding: chunked
WEBrick assumes that chunked will be set on the response using
the `#chunked=` method and doesn't respect setting the header
manually. Patch the webrick handler to set chunked on the response
in addition to setting the header. This prevents WEBrick from
calculating the content-length itself causing issues with Safari.

Fixes #618

I think this is probably a WEBrick issue which I have raised.
See: https://bugs.ruby-lang.org/issues/9986

However this patch to rack will also fix this issue for older and current
versions of Ruby (and in the future if my issue on ruby-lang continues
    to be rejected)