default_middleware_by_environment should always returns empty array for unknown keys
The current behaviour break rails server for any environment different
of development.

This behaviour was removed at 704be37e with a warning that if this
behavior is supported tests need to be added. The tests are already
there and they were not failing because "none" environment were added to
the hash. This environment is not a valid environment and it is on the
tests exactly to assert the behaviour that
default_middleware_by_environment always return an empty array for
unknown keys.