Prioritise CLI options over comment options
There are four sources of options for `Rack::Server`:

1. In the constructor, as a hash: `Rack::Server.new(server: 'thin')`
2. On the command line: `bundle exec rackup --server thin`
3. In a magic comment on the first line of config.ru (or whichever
   config file was used): `#\ --server thin`
4. `Rack::Server#default_options`

If 1 is provided, 4 and 2 won't be used. Previously, the order of
precedence - earlier 'wins' - was either:

* 1, 3
* 3, 2, 4

This changes the latter to 2, 3, 4. This allows an app to specify a
default port in the magic comment, but for that to be changed when using
the `rackup` script.