Freeze 'X-Runtime' string
Move 'X-Runtime' to a constant and freeze. Combine with other
`FORMAT_STRING` constant used in the class.

We are freezing these strings to reduce the number of allocations in
Rails integration tests. The tests are spending a lot of time in GC and
this reduces the amount of time spent from 12% to 9% (in combination
with Rails PR that also freezes some strings).

Strings allocated before this change: 1030722
Strings allocated after this change: 1024722