cache result of to_app to avoid re-traversal of maps
This prevents re-traversing the whole middleware tree and re-instantiating
Rack::Builder instances on each request.

I guess that this commit was in the same direction, except it does not solve the
problem as far as I can see:

https://github.com/rack/rack/commit/b51f3036fbdb911eac6b5cf97ebf9388fb1f14a9

This passes tests. Of course it assumes that it's fine to reuse the same
Rack::Builder instance for more than one request (I don't see why that should
not be the case).

I already use this change in a production app with a large number of mappings,
where we experienced massive improvements on routing time and object allocation.