Add support for serving gzipped static files
This makes it simple to served static files that have already been
gzipped.  This is significantly faster than using Rack::Deflater
to serve the same files, as it results in less IO and it avoids
the need to gzip at runtime.

This provides a similar feature to nginx's gzip_static directive.