Introduce a second argument for Request#parse_query
I think it's a bad idea to mix types (i.e. String and Hash here),
and Utils.parse_query already accepts a second argument for
passing the delimiter, so why not do the same?

Ref: https://github.com/rack/rack/pull/899

My motivation for patching this is because we have overridden
Request#parse_query as follows, and it broke due to the above
pull request:

    class Request < Rack::Request
      def parse_query(qs)
        Rack::Utils.parse_query(qs, '&')
      end
    end

Given my patch, our application would still break, but at least
we just need to accept the delimiter and pass it to Utils.parse_query,
rather than doing String/Hash handling again.

I accepting this patch means our application would break again,
but I'll just adopt the change again when a new Rack version is
released.

By the way, the intention for overriding Request#parse_query was
trying to avoid Utils.parse_nested_query because I hate that
behaviour. If this could be an option for Rack::Request, that
would be awesome.

If we're willing to make this change, then I'll try to think about
how to make it an option.