Convenient cache and content type for response generation
These three convenient methods make it easy to declaratively specify how a response should be structured. For example:

    response.do_not_cache!

    response.cache!

    response.content_type! 'text/html'

Open to discussion w.r.t. naming, and behaviour. I monkey patch these in which might be fine for my usage, but I thought it could be useful to a wider audience.