fix rdoc formatting (remove accidental links)
In several places the rdoc comments attempt to indicate hash access, for
example `env['rack.input']`, but rdoc interprets this as a link with
text "env" and href="%27rack.input%27/". Wrapping these snippets of hash
access code in `<tt>` tags escapes the link syntax and also renders the
code snippets in monospace.