Support RFC 7239: HTTP Forwarded header
[RFC 7239][] is an IETF proposed standard for a HTTP `Forwarded` request
header that aims to replace non-standard `X-Forwarded-*` headers.

This commit supports all relevant tokens that may be set in the
Forwarded header:
  - proto
  - host
  - for

I have prioritised the Forwarded header over the non-standard headers
as I think it's preferable to use the standard header when specified.

Multiple values for each token are allowed by the standard. When
multiple values are supplied, I've followed these rules:
  - for the `proto` token, use the first value specified as it's most
    likely to be the protocol used by the original client
  - for the `host` token, use the last value specified following the
    convention already used by Rack for multiple `X-Forwarded-For`
    headers

When both `X-Forwarded-For` and `Forwarded` headers are supplied and the
latter includes `for` values, I've concatenated both sets of values with
the `Forwarded` values coming last.

I've been careful to support mixed case in the header and optional
double quotes encapsulating the token value as specified in the RFC.

[RFC 7239]: https://tools.ietf.org/html/rfc7239