Update cookie parser to only accept `;` as delimiter.
The cookie parse used to accept both `;` and `,` as delimiters. This
follows the original specifications in RFC 2109 and RFC 2965, but both of
these have been obsoleted by RFC 6265 which does not include the
possibility of `,` as a delimiter in Cookie: headers. Note that `,` is
still not allowed in cookie values (it is explicitly excluded from
cookie-octets in RFC 6265).

From now on the cookie parser will only accept `;` as a delimiter to
adhere to RFC 6265.