deflater: remove "deflate" encoding support
This will improve cache hit rates and reduce caching overhead at
small expense of increased header overhead for some user agents.
For reference, Varnish cache supports only gzip as well:

	https://www.varnish-cache.org/docs/4.1/phk/gzip.html

In the past, "deflate" encoding was more likely to trigger
user-agent bugs, as noted in the comments removed with this
change as well as the Varnish documentation referenced above.