Fix handling of parsing empty nested queries
If the query string is empty, instead of using a plain hash, we
still need to use params.to_params_hash, as params.to_params_hash
may return a different object, such a hash with a default proc.

This fixes request.params in cases where indifferent query
params are used, and r.GET is empty but r.POST is not.  Before
this commit, this would result in request.params ending up with
a non-indifferent hash, when it should have used an indifferent
hash.

The cause of this issue can originally be traced to an optimization
in c4596b3e3d821d19b25c2d48d619097a3ad39ec0, which was before I
added the code to allow user-configurable query params in
7e7a3890449b5cf5b86929c79373506e5f1909fb.  Now that
user-configurable query params are supported, the optimization is
invalid and needs to be removed.