Use attr reader when referencing cookie key
Allows custom session stores that extend the abstract session store to
dynamically select which cookie key will be used for fetching/setting
the session's id. This allows the user to maintain multiple, different sessions.
The web application can determine which session to be used for the
current request based on any number of conditions (URL, headers, cookies, IP,
etc).

Fixes #1155