Don't fall back to REMOTE_ADDR unless X_FORWARDED_FOR is empty
The previous, faulty behaviour was to remove all trusted ip addresses
from X_FORWARDED_FOR and returning the last one remaining, falling back
to the REMOTE_ADDR if none remain.

A problem arises if all the address in X_FORWARDED_FOR are trusted.
They can't all be proxies. One of them (the first one) is the ip that
made the request. The correct behaviour is to return the first one, but
the previous implementation would remove them all, and fall back to
REMOTE_ADDR.

This patch removes trusted ips only from the tail of X_FORWARDED_FOR,
and returns the last address in the resulting list, ensuring that we
don't fall back to REMOTE_ADDR if X_FORWARDED_FOR is non-empty.