Don't use #eof? when parsing multipart
On chunked requests (`Transfer-Encoding: chunked`) content length won't
be present, so `#eof?` will be called on the Rack input directly.
However, the Rack input doesn't have to respond to `#eof?`, it's not
part of the Rack specification. For example, `Rack::Lint::InputWrapper`
and `Unicorn::StreamInput` don't respond to `#eof?`. This means that in
development and on Unicorn, multipart parsing will fail for chunked
request.

This change refactors the multipart parser so that it doesn't call
`#eof?`.