According to RFC 6265 [1]:
> To maximize compatibility with user agents, servers that wish to
store arbitrary data in a cookie-value SHOULD encode that data, for
example, using Base64 [RFC4648]

RFC 4648 uses m0 for the encoding (according to String#unpack in the
Ruby docs [2]). It's a SHOULD not a MUST but I don't see anything in
the logs to indicate why the advice isn't followed.

The encoding was introduced in 417ac6a and refactored in e6b6529.

[1] https://tools.ietf.org/html/rfc6265#section-4.1.1
[2] https://ruby-doc.org/core-2.6.3/String.html#method-i-unpack