Use `const_get` to get klass without `inject`
`Enumerable#inject` was used to get constant in `Rack::Handler.get`.
But `Object#const_get` is available to get nested constant from Ruby version 2.0.0.
And the required minimum ruby version of Rack is 2.2.2.
It seems that `const_get` doesn't need `inject` anymore,
and this change is for replacing it without `inject`.