Make SessionId#to_s be an alias of #public_id
This restores backwards compatibility with code that uses the
session id.

No justification was given for making to_s raise in the commit
that changed the behavior (cc1d162d28396b6a71f266e6a40ffc19a258792b).

If there is a reason to keep raising an exception, a specific
exception class and explicit error message should be used.