Make body not respond to to_path to Rack::Files range requests
If the body responds to to_path, the server is allowed to serve
the file at that path and ignore the body.  That works well for
most requests, but not for range requests, since the correct
byte range won't be served.

Work around this issue by using a separate body class for the
partial content responses that does not respond to to_path, and
the current body class for non-partial content responses.

Fixes #1235