Fix incorrect MockResponse#body String encoding
This is a regression introduced in Rack 2.1.0, possibly after
8c62821f4a464858a6b6ca3c3966ec308d2bb53e was introduced.

Before the aforementioned commit, `MockResponse#body` would return a
UTF-8 string as it joins all the response parts together. However,
because the aforementioned commit uses `String.new` without specifying
an encoding, Ruby would create a new String with `ASCII-8BIT` encoding
by default[1].

This commit adds failing test cases to prevent future regression to
ensure that we return UTF-8 String for all string body, and also making
sure that we return binary string if any of the response part is a
binary string (with `ASCII-8BIT` encoding).

[1]: https://ruby-doc.org/core-2.7.0/String.html#method-c-new