Update SPEC and Lint to require ASCII-8BIT encoding for CGI values with non-ASCII characters
Fixes warning when binary Regexp is used to match against a
CGI value (QUERY_STRING).

ASCII-8BIT encoding is already required for response bodies.
Requiring this encoding for all CGI values would break usage
with puma and potentially other rack servers. So it seems
safest to only require ASCII-8BIT encoding when the value
actually contains non-ASCII.  We could further limit it to
just QUERY_STRING, but I'm not sure we want to do that.

Testing with a few rack servers:

* Unicorn always uses ASCII-8BIT for all CGI values.
* Puma will use UTF-8 for some values, but uses ASCII-8BIT
  for CGI values contianing non-ASCII characters.
* Webrick doesn't support non-ASCII chracters, at least for
  PATH_INFO or QUERY_STRING.

If we don't do this, the only way to avoid the warning would be
to duplicate QUERY_STRING if it is not already binary.  Checking
encoding will slow down parsing even if it already uses
ASCII-8BIT encoding.

Fixes #1183