Support providing IO-like object to Rack::Multipart::UploadedFile
This changes the UploadedFile API to use keyword arguments, being
backwards compatible with the previous API by keeping the positional
arguments and making the keyword defaults be the positional argument
values.

The two new arguments to UploadedFile#initialize are filename and io.
filename allows overriding the filename that will be used, instead of
always using the base name of the path.  io specifies an IO-like
object that can be read from to get the data, overridding any
provided path.  This allows you to generate UploadedFile instances
that previously were not possible without accessing the file system.

This changes the multipart generator to only include a Content-Length
for files with paths, and only include a filename in
Content-Disposition if the UploadedFile has one.

Fixes #1028.