Use require_relative and rely on autoload instead of using require
This does a better job of testing the autoloads work, and ensures
files are loaded from the current rack directory.

Some files required rack.rb, and in those cases I didn't remove
the require, just replaced it with require_relative, since the
intent is to make sure all of rack is loaded if a file under
rack is loaded.

This has the potential to break code that does:

  require 'rack/content_type'
  Rack::ContentType.new(app)

If the rack library itself is never required.  This is because
the autoloads that rack.rb sets up would not be loaded. I'm not
sure if we want to support that, but if so we would want to keep
the requires and convert them to require_relative.

For core_ext/regexp, move the requires into the if < 2.4 block,
so that the file isn't loaded unless it is necessary.