Fix Response#write calculation of Content-Length if initialized with a body
Before, @length was always set to 0 in the constructor. The breaks
the case where a string was provided, since that is automatically
bufferred.  Fix that by using the bytesize of the string.

If body was provided as an array, when switching to buffered mode,
iterate over the body to figure out the content length.

Note this only occurs when doing Response#write, as that is the
only case where Response updates the Content-Length.

I found one spec that called Response#write inside a Response#finish
block.  That updates the response body after the header hash is
returned, and does not call Response#append as the @writer has been
changed to the Response#finish block, so it doesn't update the
Content-Length in the returned header hash.