Remove dead code in multipart parser
This can only be hit if filename is nil/false.  However, when
the MimePart is created, it is either a TempfilePart if there
is a filename or a BufferPart otherwise.  In the TempfilePart
case, the filename is present so this code is not hit.  In the
BufferPart case, the body is a String and not an IO, so this
cannot be hit.