Make Lint check response hijacking
The previous code to do this was broken, since it modified a
local variable and not the actual response headers.  Note that
this checking can only be done if the response headers are a
hash, as this requires modifying the response headers, so it
is only done in that case.