Refactor and document Chunked
The most important part of the documentation is how to use the
Trailer header correctly.  It was not obvious before that the
response body is expected to have a trailers method already
defined.

Body doesn't need to include Rack::Utils.  This was originally done
for the bytesize method, which is no longer used.

Rename insert_trailers to yield_trailers to reflect what the method
does.

Other minor refactorings for better performance and simpler code.