Refactor and document Deflater
Explicit :sync => nil option is now treated as false instead of true.
We documented the default was true, but any explicit :sync option
should be respected.

gzip.write returns the length of the input string, so simplify the
GzipStream#each implementation by skipping all empty body strings.

Don't set @writer and @body to nil in GzipStream after using them.
We don't use this pattern in BodyProxy, so I don't see a need to
use it here.