Refactor and document Directory
The previous implementation of directory listings was inefficient for large
directories.  First it built a bunch of file entry strings, then it joined
them into a large string containing all entries, then that was interpolated
into the header and footer creating another large string.  Then each_line was
called on the large string, creating one string per line.

This switches to a streaming approach for directory listings.  We yield the
header, then one string for each listing, then the footer.

Testing with puma with a directory with 10,000 files with 100 byte file
names, directory listing before took 3.5 seconds, with 900ms before
first byte.  After this commit, it takes 2.1 seconds, with 10ms before
the first byte.  Note that total time is not hugely different, because
most of the time is probably taken in stat(2) calls.  Time to first byte is
greatly improved, though.

Remove Directory#path.  path is used as a local variable, but the instance
variable is never set, so the value would always be nil.

The previous approach used a glob for getting files.  That resulted in
broken behavior when the path being served includes glob metacharacters.
Switch to a `Dir.foreach` approach to avoid that issue.  Because the glob
skips hidden files, continue to skip files starting with a period.  Note
that Directory will still serve hidden files and display directory listings
for hidden directories, so this should not be considered a security feature.

As a small behavior change, do not show a Parent directory link for the
root directory, since it takes you to the same page.

Switch list_path to not use exceptions for flow control.  Remove all
exception handling, and switch to using Directory#stat, which already handles
the same exceptions.  This has the advantage so that if a non-file,
non-directory is requested, you get a valid 404 rack response instead of
nil.