Ensure that multiline input doesn't cause Rack::Request to break
As written, the `AUTHORITY` regex intends to match the entire input string, and the sole usage of it expects that it will always return a match.  After some consideration, there *is* a way to feed it input that will not match, causing errors on the subsequent line.

The issue at hand is that `.` in a regular expression matches everything *except newlines* by default, meaning that even the most generous regular expression with start- and end-of-string anchors (`/\A.*\z/`) cannot match a multiline string.  The solution, then, is to use the `m` (multiline) flag on the regex, which allows `.` to match newlines as well.