Rollback support for newlines in hostnames
If the address cannot reasonably be used to reconstruct a URI, we return nil.  This includes generous capture patterns for IPv6 and IPv4 addresses, with validation of the captured data happening as a secondary pass.

Additionally added some testing with extreme examples of valid inputs, and a few invalid ones.