Fix possible thread safety issue in Rack::Session::Pool
Internal use in Rack::Session::Pool is already thread-safe.
However, this can be called from commit_session and is a public
method, so make those calls thread-safe.

Note that Rack::Session::Pool#generate_sid did no support taking
the same arguments as the superclass #generate_sid did.  This
overrides the method to support the same arguments, but adds a
keyword argument for whether to use a mutex, which defaults to
true.