Explicitly require each rack dependency of file
The goal here is to enable a consumer of Rack to load only a subset of
Rack (e.g., `require "rack/request"`) without having to pull in all of
Rack via `require "rack"`. This is a common pattern for tools build to
work atop or next to Rack. A prior change
(0d14bf5797ebe9ef51f55102e27b9c147c1dc283) was made to lean more heavily
on Ruby's `autoload` feature, breaking this usage pattern. This commit
aims to re-introduce `require_relative`s necessary to allow any subset
of Rack to be directly loaded.

In some cases, that means a file must also effectively `require_relative
'lib/rack'` because the file has dependencies on constants defined in
`lib/rack.rb`. The net effect is that no other `require_relatives` would
then be needed in such a file b/c the `autoload` configuration of
`lib/rack.rb` would be in play. But we've chosen to still
`require_relative` those dependencies as it helps to highlight the
dependencies used within the file.